+++ src/main.rs
@@ -3,7 +3,7 @@
  * License: MIT
  */
 
-use anyhow::Result;
+use anyhow::{Result, bail};
 use flume::{Receiver, Sender};
 use simple_logger::SimpleLogger;
 use std::path::PathBuf;
@@ -25,7 +25,7 @@ mod tagreader;
 use crate::client::Client;
 use crate::config::PartyConfig;
 use crate::events::Event;
-use crate::model::UserId;
+use crate::model::{UserId, UserMode};
 
 fn main() -> Result<()> {
     SimpleLogger::new()
@@ -56,6 +56,12 @@ fn main() -> Result<()> {
 fn run(config_filename: PathBuf) -> Result<()> {
     let config = config::load_config(&config_filename)?;
 
+    let user_mode = config.get_user_mode();
+    match &user_mode {
+        UserMode::SingleUser(id) => log::info!("Running in single-user mode for user ID '{id}'."),
+        UserMode::MultiUser => log::info!("Running in multi-user mode."),
+    }
+
     let sounds_path = config.sounds_path.clone();
 
     let (tx1, rx): (Sender<Event>, Receiver<Event>) = flume::unbounded();
@@ -64,7 +70,13 @@ fn run(config_filename: PathBuf) -> Result<()> {
 
     ctrlc::set_handler(move || handle_ctrl_c(&tx1)).expect("Could not set Ctrl-C handler");
 
-    tagreader::handle_tag_reads(config.reader_input_device, tx2)?;
+    if let UserMode::MultiUser = user_mode {
+        match config.reader_input_device {
+            Some(device) => tagreader::handle_tag_reads(device, tx2)?,
+            None => bail!("No reader device configured, but one is required in multi-user mode."),
+        }
+    }
+
     buttons::handle_button_presses(
         config.button_input_device,
         config.buttons_to_key_code_names,
@@ -75,7 +87,7 @@ fn run(config_filename: PathBuf) -> Result<()> {
 
     client.sign_on()?;
 
-    handle_events(rx, client, &config.party)?;
+    handle_events(rx, client, &config.party, &user_mode)?;
 
     Ok(())
 }
@@ -84,6 +96,7 @@ fn handle_events(
     event_receiver: Receiver<Event>,
     mut client: Client,
     party_config: &PartyConfig,
+    user_mode: &UserMode,
 ) -> Result<()> {
     let mut current_user_id: Option<UserId> = None;
 
@@ -95,15 +108,27 @@ fn handle_events(
             }
             Event::ButtonPressed { button } => {
                 log::debug!("Button pressed: {:?}", button);
-                // Submit if user has identified; ignore if no user has
-                // been specified.
-                if let Some(user_id) = current_user_id {
-                    client.handle_button_press_with_identified_user(
-                        user_id,
-                        button,
-                        party_config,
-                    )?;
-                    current_user_id = None; // reset
+
+                match user_mode {
+                    UserMode::SingleUser(user_id) => {
+                        client.handle_button_press_with_identified_user(
+                            user_id.clone(),
+                            button,
+                            party_config,
+                        )?;
+                    }
+                    UserMode::MultiUser => {
+                        // Submit if user has identified; ignore if no user has
+                        // been specified.
+                        if let Some(user_id) = current_user_id {
+                            client.handle_button_press_with_identified_user(
+                                user_id,
+                                button,
+                                party_config,
+                            )?;
+                            current_user_id = None; // reset
+                        }
+                    }
                 }
             }
             Event::ShutdownRequested => {